+++ app/src/main/java/com/daniil/halushka/todoapp/presentation/screens/elements/CustomTopBar.kt
@@ -1,0 +1,79 @@
+package com.daniil.halushka.todoapp.presentation.screens.elements
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.daniil.halushka.todoapp.R
+import com.daniil.halushka.todoapp.data.repository.TodoRepository
+import com.daniil.halushka.todoapp.ui.theme.TodoAppTheme
+
+@Composable
+fun CustomTopBar(
+    completedItemsCount: Int,
+) {
+    Box(
+        modifier = Modifier
+            .background(MaterialTheme.colorScheme.primary),
+    ) {
+        Surface(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 8.dp)
+                .height(64.dp)
+        ) {
+            Row(
+                modifier = Modifier
+                    .background(MaterialTheme.colorScheme.primary),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(
+                    modifier = Modifier
+                        .weight(1f),
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = stringResource(R.string.my_tasks_todo_s),
+                        fontSize = 24.sp,
+                        fontWeight = FontWeight.Bold,
+                        color = MaterialTheme.colorScheme.onBackground
+                    )
+                    Text(
+                        text = stringResource(
+                            id = R.string.completed_todo_s,
+                            completedItemsCount
+                        ),
+                        fontSize = 20.sp,
+                        color = MaterialTheme.colorScheme.onTertiary
+                    )
+                }
+            }
+        }
+    }
+}
+
+
+@Preview(showBackground = true)
+@Composable
+fun Preview1() {
+    TodoAppTheme {
+        Column {
+            CustomTopBar(completedItemsCount = 5)
+            ContainerWithTodo(repository = TodoRepository()) {}
+        }
+    }
+}
