+++ app/src/main/java/com/daniil/halushka/todoapp/ui/theme/Color.kt
@@ -41,12 +41,3 @@ val nightLabelDisableColor = Color(0x26FFFFFF)
 val nightBackPrimaryColor = Color(0xFF161618)
 val nightBackSecondaryColor = Color(0xFF252528)
 val nightBackElevatedColor = Color(0xFF3C3C3F)
-
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
